added filter for candidate type
